+++ src/app/pages/game-gallery/game-gallery.component.html
@@ -34,10 +34,10 @@
 
         <card-info-game
         title="Ajedrez"
-        description="Desafia a la IA"
+        description="Juega contra un usuario o la IA"
         buttonText="Jugar"
         imageUrl="assets/img/gameGallery/juego5.png"
-        routerLink="/chess"
+      routerLink="/chess"
       ></card-info-game>
      
       <card-info-game
