+++ src/app/pages/games/chess/chess-board/chess-board.service.ts
@@ -1,0 +1,10 @@
+import { Injectable } from '@angular/core';
+import { BehaviorSubject } from 'rxjs';
+import { FENConverter } from '../chess-logic/FENConverter';
+
+@Injectable({
+  providedIn: 'root'
+})
+export class ChessBoardService {
+  public chessBoardState$ = new BehaviorSubject<string>(FENConverter.initalPosition);
+}
