+++ src/app/pages/games/chess/chess-board/models.ts
@@ -1,0 +1,15 @@
+import { FENChar } from "../chess-logic/models"
+
+type SquareWithPiece = {
+    piece: FENChar;
+    x: number;
+    y: number;
+}
+
+type SquareWithoutPiece = {
+    piece: null;
+}
+
+export type SelectedSquare = SquareWithPiece | SquareWithoutPiece;
+
+// columns are defined in chess-logic/models to be shared across app
