+++ src/app/pages/games/chess/chess-timer/chess-timer.component.css
@@ -1,0 +1,2 @@
+.timer { display: inline-flex; align-items: center; gap: .5rem; font-family: system-ui, Arial, sans-serif; }
+.time { font-weight: 600; font-variant-numeric: tabular-nums; }
