+++ src/app/pages/games/chess/chess-timer/chess-timer.component.html
@@ -1,0 +1,7 @@
+<div class="timer">
+  <span class="time">{{ display }}</span>
+  <!-- Controles opcionales para pruebas locales (no usar en producción si no se necesitan) -->
+  <!-- <button (click)="start()">Start</button>
+  <button (click)="stop()">Stop</button>
+  <button (click)="reset()">Reset</button> -->
+</div>
