+++ src/app/pages/games/chess/modules/chess-board/chess-board.service.ts
@@ -1,11 +1,1 @@
-import { Injectable } from '@angular/core';
-import { BehaviorSubject } from 'rxjs';
-import { FENConverter } from '../../chess-logic/FENConverter';
-
-
-@Injectable({
-  providedIn: 'root'
-})
-export class ChessBoardService {
-  public chessBoardState$ = new BehaviorSubject<string>(FENConverter.initalPosition);
-}
+export * from '../../chess-board/chess-board.service';
