+++ src/app/pages/games/chess/modules/chess-board/models.ts
@@ -1,15 +1,1 @@
-import { FENChar } from "../../chess-logic/models"
-
-type SquareWithPiece = {
-    piece: FENChar;
-    x: number;
-    y: number;
-}
-
-type SquareWithoutPiece = {
-    piece: null;
-}
-
-export type SelectedSquare = SquareWithPiece | SquareWithoutPiece;
-
-export const columns = ["a", "b", "c", "d", "e", "f", "g", "h"] as const;
+export * from '../../chess-board/models';
