--- src/app/pages/games/chess/modules/chess-timer/chess-timer.component.html
@@ -1,4 +1,0 @@
-<div class="timer-container">
-  <span class="timer-label">⏱</span>
-  <span class="timer-value">{{ formattedTime }}</span>
-</div>
