--- src/app/pages/games/chess/modules/choose-mode-dialog/choose-mode-dialog.component.css
@@ -1,74 +1,0 @@
-
-/* Unificado con el modal de la IA */
-.custom-dialog-content {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  padding: 2rem 2.5rem;
-  background: #fff;
-  border-radius: 18px;
-  box-shadow: 0 8px 32px rgba(0,0,0,0.18);
-  min-width: 500px;
-  max-width: 650px;
-}
-.dialog-title {
-  color: #23408e;
-  font-weight: 600;
-  margin-bottom: 1.5rem;
-  font-size: 1.5rem;
-  text-align: center;
-}
-.choose-mode-options {
-  width: 100%;
-  display: flex;
-  flex-direction: row;
-  gap: 2rem;
-  justify-content: center;
-  margin-top: 1.2rem;
-}
-.choose-btn {
-  min-width: 210px;
-  min-height: 64px;
-  font-weight: 700;
-  font-size: 1.25rem;
-  background: linear-gradient(90deg, #23408e 60%, #3a5db7 100%);
-  color: #fff !important;
-  border-radius: 18px;
-  box-shadow: 0 4px 18px 0 rgba(35,64,142,0.13);
-  padding: 0.9rem 0.5rem 0.9rem 0.5rem;
-  letter-spacing: 0.5px;
-  border: none;
-  outline: none;
-  transition: background 0.18s, box-shadow 0.18s, transform 0.12s;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  gap: 0.7rem;
-}
-.choose-btn:hover {
-  background: linear-gradient(90deg, #19306a 60%, #23408e 100%);
-  box-shadow: 0 8px 24px 0 rgba(35,64,142,0.18);
-  transform: translateY(-2px) scale(1.03);
-}
-.choose-icon {
-  font-size: 2rem;
-  margin-right: 0.1rem;
-}
-/* El bloque anterior era inválido y se elimina para corregir la sintaxis CSS */
-@media (max-width: 500px) {
-  .custom-dialog-content {
-    min-width: 90vw;
-    max-width: 98vw;
-    padding: 1.2rem 0.5rem 1.2rem 0.5rem;
-  }
-  .menu-btns {
-    flex-direction: column;
-    gap: 1.2rem;
-  }
-  button[mat-raised-button] {
-    min-width: 0;
-    width: 100%;
-    font-size: 1.05rem;
-    padding: 0.8rem 0;
-  }
-}
