--- src/app/pages/games/chess/modules/choose-mode-dialog/choose-mode-dialog.component.html
@@ -1,13 +1,0 @@
-<div class="custom-dialog-content">
-  <h2 class="dialog-title">¿Cómo quieres jugar?</h2>
-  <div class="choose-mode-options">
-    <button mat-raised-button class="choose-btn" (click)="chooseFriend()">
-      <span class="choose-icon">👤</span>
-      Jugar contra usuario
-    </button>
-    <button mat-raised-button class="choose-btn" (click)="chooseAI()">
-      <span class="choose-icon">🤖</span>
-      Jugar contra IA
-    </button>
-  </div>
-</div>
