--- src/app/pages/games/chess/modules/computer-mode/computer-mode.component.html
@@ -1,42 +1,0 @@
-
-<ng-container *ngTemplateOutlet="chessBoard"></ng-container>
-
-<ng-template #chessBoard>
-  <!-- Copia el contenido original de chess-board.component.html aquí, sin la cabecera de Jugador -->
-  <div class="chess-board-container">
-      <app-chess-timer *ngIf="timerStarted" style="margin-bottom: 10px;"></app-chess-timer>
-    <div class="chess-board" [ngClass]="{'rotated': flipMode}">
-        <div *ngFor="let row of chessBoardView; let x = index" class="row">
-            <div *ngFor="let piece of row; let y = index" class="square" [ngClass]="{
-                'dark': isSquareDark(x, y), 
-                'light': !isSquareDark(x, y),
-                'selected-square': isSquareSelected(x, y),
-                'last-move': isSquareLastMove(x, y),
-                'king-in-check': isSquareChecked(x, y),
-                'promotion-square': isSquarePromotionSquare(x, y)
-            }" (click)="move(x, y)">
-                <div [ngClass]="{'safe-square': isSquareSafeForSelectedPiece(x, y)}"></div>
-                <img *ngIf="piece" [src]="getPieceImage(piece)" [alt]="piece" class="piece"
-                    [ngClass]="{'rotated': flipMode}">
-            </div>
-        </div>
-    </div>
-    <!-- Botón Flip eliminado para la vista de la IA -->
-      <!-- Botón volver solo visible en FriendModeComponent, no aquí -->
-          <div class="move-list-wrapper">
-              <app-move-list [moveList]="moveList" [gameHistoryPointer]="gameHistoryPointer" [gameHistoryLength]="gameHistory.length"
-                      (showPreviousPositionEvent)="showPreviousPosition($event)"></app-move-list>
-          </div>
-          <h2 *ngIf="gameOverMessage" class="game-over-message">
-                  {{gameOverMessage}}
-          </h2>
-  </div>
-
-  <div *ngIf="isPromotionActive" class="promotion-dialog">
-      <img *ngFor="let piece of promotionPieces()" [src]="getPieceImage(piece)" (click)="promotePiece(piece)">
-
-      <span class="close-promotion-dialog" (click)="closePawnPromotionDialog()">
-          &times;
-      </span>
-  </div>
-</ng-template>
