+++ src/app/pages/games/chess/modules/computer-mode/stockfish.service.ts
@@ -1,71 +1,1 @@
-import { HttpClient, HttpParams } from '@angular/common/http';
-import { Injectable } from '@angular/core';
-import { ChessMove, ComputerConfiguration, StockfishQueryParams, StockfishResponse, stockfishLevels } from './models';
-import { BehaviorSubject, Observable, of, switchMap } from 'rxjs';
-import { Color, FENChar } from '../../chess-logic/models';
-
-@Injectable({
-  providedIn: 'root'
-})
-export class StockfishService {
-  private readonly api: string = "https://stockfish.online/api/s/v2.php";
-
-  public computerConfiguration$ = new BehaviorSubject<ComputerConfiguration>({ color: Color.Black, level: 1 });
-
-  constructor(private http: HttpClient) { }
-
-  private convertColumnLetterToYCoord(string: string): number {
-    return string.charCodeAt(0) - "a".charCodeAt(0);
-  }
-
-  private promotedPiece(piece: string | undefined): FENChar | null {
-    if (!piece) return null;
-    const computerColor: Color = this.computerConfiguration$.value.color;
-    if (piece === "n") return computerColor === Color.White ? FENChar.WhiteKnight : FENChar.BlackKnight;
-    if (piece === "b") return computerColor === Color.White ? FENChar.WhiteBishop : FENChar.BlackBishop;
-    if (piece === "r") return computerColor === Color.White ? FENChar.WhiteRook : FENChar.BlackRook;
-    return computerColor === Color.White ? FENChar.WhiteQueen : FENChar.BlackQueen;
-  }
-
-  private moveFromStockfishString(move: string): ChessMove {
-    const prevY: number = this.convertColumnLetterToYCoord(move[0]);
-    const prevX: number = Number(move[1]) - 1;
-    const newY: number = this.convertColumnLetterToYCoord(move[2]);
-    const newX: number = Number(move[3]) - 1;
-    const promotedPiece = this.promotedPiece(move[4]);
-    return { prevX, prevY, newX, newY, promotedPiece };
-  }
-
-  public getBestMove(fen: string): Observable<ChessMove> {
-    const queryParams: StockfishQueryParams = {
-      fen,
-      depth: stockfishLevels[this.computerConfiguration$.value.level],
-    };
-
-    let params = new HttpParams().appendAll(queryParams);
-
-    return this.http.get<StockfishResponse>(this.api, { params })
-      .pipe(
-        switchMap(response => {
-          const bestMove: string = response.bestmove.split(" ")[1];
-          return of(this.moveFromStockfishString(bestMove));
-        })
-      )
-  }
-
-  // Método directo para hacer que la IA juegue - más confiable para el movimiento inicial
-  public play(fen: string): void {
-    console.log('StockfishService: Solicitando movimiento para FEN: ' + fen);
-    this.getBestMove(fen).subscribe({
-      next: (move) => {
-        console.log('StockfishService: Movimiento recibido:', move);
-        // Creamos un evento personalizado de forma compatible con TypeScript
-        const event = new CustomEvent('stockfishMove', { detail: move }) as Event;
-        window.dispatchEvent(event);
-      },
-      error: (err) => {
-        console.error('Error al obtener movimiento de Stockfish:', err);
-      }
-    });
-  }
-}
+export * from '../../computer-mode/stockfish.service';
