--- src/app/pages/games/chess/modules/friend-mode/index.ts
@@ -1,1 +1,0 @@
-export * from './friend-mode.component';
