--- src/app/pages/games/chess/modules/nav-menu/nav-menu.component.html
@@ -1,1 +1,0 @@
-<router-outlet></router-outlet>
