--- src/app/pages/games/chess/modules/play-against-computer-dialog/play-against-computer-dialog.component.css
@@ -1,72 +1,0 @@
-/* Unificado y corregido: */
-.custom-dialog-content {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    padding: 2rem 2.5rem;
-    background: #fff;
-    border-radius: 18px;
-    box-shadow: 0 8px 32px rgba(0,0,0,0.18);
-    min-width: 500px;
-    max-width: 650px;
-}
-.dialog-title {
-    color: #23408e;
-    font-weight: 600;
-    margin-bottom: 1.5rem;
-}
-.ia-options {
-    width: 100%;
-    margin-top: 1.5rem;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-}
-.levels-title {
-    font-size: 1.1rem;
-    color: #23408e;
-    margin-bottom: 0.5rem;
-}
-.stockfish-strength-container {
-    display: flex;
-    gap: 1rem;
-    margin-bottom: 1.2rem;
-}
-.strength-div {
-    border: 2px solid #23408e;
-    border-radius: 8px;
-    padding: 0.4rem 1.2rem;
-    font-weight: 500;
-    color: #23408e;
-    cursor: pointer;
-    background: #fff;
-    transition: background 0.2s, color 0.2s;
-}
-.strength-div.selected {
-    background: #23408e;
-    color: #fff;
-}
-.choose-side {
-    display: flex;
-    gap: 2.5rem;
-    margin: 1.2rem 0 0.2rem 0;
-}
-.king-img {
-    width: 48px;
-    height: 48px;
-    cursor: pointer;
-    border-radius: 50%;
-    border: 2px solid transparent;
-    transition: border 0.2s;
-}
-.king-img:hover {
-    border: 2px solid #23408e;
-}
-.side-labels {
-    display: flex;
-    justify-content: space-between;
-    width: 100%;
-    padding: 0 2.5rem;
-    color: #23408e;
-    font-size: 1rem;
-}
