--- src/app/pages/games/chess/modules/play-against-computer-dialog/play-against-computer-dialog.component.html
@@ -1,17 +1,0 @@
-<div class="custom-dialog-content">
-    <h2 class="dialog-title">Configura tu partida contra la IA</h2>
-    <div class="ia-options">
-        <div class="levels-title">Dificultad de la IA:</div>
-        <div class="stockfish-strength-container">
-            <div *ngFor="let level of levels" class="strength-div" (click)="selectStockfishLevel(level.value)"
-                [ngClass]="{'selected': level.value === stockfishLevel}">
-                {{level.label}}
-            </div>
-        </div>
-        <div class="choose-side">
-            <img src="assets/img/chesspieces/white king.svg" class="king-img" draggable="false" (click)="play('w')">
-            <img src="assets/img/chesspieces/black king.svg" class="king-img" draggable="false" (click)="play('b')">
-        </div>
-    <!-- Etiquetas de lado removidas, solo imágenes -->
-    </div>
-</div>
