+++ src/app/pages/games/chess/move-list/move-list.component.css
@@ -1,0 +1,44 @@
+.move-list {
+    max-height: 100px;
+    overflow-y: scroll;
+    display: flex;
+    flex-direction: column-reverse;
+}
+
+.move-list::-webkit-scrollbar {
+    display: none;
+}
+
+.move-list {
+    -ms-overflow-style: none;
+    scrollbar-width: none;
+}
+
+.row {
+    width: 250px;
+    display: flex;
+    color: white
+}
+
+.move-number {
+    width: 50px;
+}
+
+.current-move {
+    background-color: #0d47a1;
+}
+
+.move {
+    cursor: pointer;
+    width: 100px;
+}
+
+.move:hover {
+    background-color: #0d47a1;
+    opacity: 0.7
+}
+
+/* Material icon buttons to primary blue */
+button[mat-icon-button] {
+    color: #0d47a1;
+}
