+++ src/app/pages/games/chess/move-list/move-list.component.html
@@ -1,4 +1,8 @@
 <div>
+    <button mat-icon-button color="primary" [disabled]="gameHistoryPointer === 0" (click)="showPreviousPosition(0)">
+        <mat-icon>first_page</mat-icon>
+    </button>
+
     <button mat-icon-button color="primary" [disabled]="gameHistoryPointer === 0"
         (click)="showPreviousPosition(gameHistoryPointer - 1)">
         <mat-icon>navigate_before</mat-icon>
@@ -7,6 +11,11 @@
     <button mat-icon-button color="primary" [disabled]="gameHistoryPointer === gameHistoryLength - 1"
         (click)="showPreviousPosition(gameHistoryPointer + 1)">
         <mat-icon>navigate_next</mat-icon>
+    </button>
+
+    <button mat-icon-button color="primary" [disabled]="gameHistoryPointer === gameHistoryLength - 1"
+        (click)="showPreviousPosition(gameHistoryLength - 1)">
+        <mat-icon>last_page</mat-icon>
     </button>
 </div>
