+++ src/app/pages/games/chess/nav-menu/nav-menu.component.html
@@ -1,0 +1,12 @@
+<mat-toolbar>
+    <button mat-button routerLink="/against-friend">
+        Play against friend
+    </button>
+
+    <button mat-button (click)="playAgainstComputer()">
+        Play against computer
+    </button>
+
+</mat-toolbar>
+
+<router-outlet></router-outlet>
