+++ src/app/pages/games/crossword-game/crossword-game.component.html
@@ -1,89 +1,4 @@
-<app-nav></app-nav>
-
-<div class="crossword-container">
-  <div class="crossword-header">
-    <div class="difficulty-selector" [ngClass]="difficulty">
-      <label for="difficulty">Dificultad:</label>
-      <select id="difficulty" [(ngModel)]="difficulty" (change)="onDifficultyChange()">
-        <option value="easy">Fácil</option>
-        <option value="medium">Medio</option>
-        <option value="hard">Difícil</option>
-      </select>
-    </div>
-
-    <div class="timer">Tiempo: {{ timer }}</div>
-  </div>
-
-  <div class="crossword-content">
-    <!-- Rejilla -->
-    <div class="grid-wrapper">
-      <table class="crossword-grid">
-        <tr *ngFor="let row of grid; let ri = index">
-          <td *ngFor="let cell of row; let ci = index">
-            <ng-container *ngIf="cell === null; else editTpl">
-              <div class="black-cell"></div>
-            </ng-container>
-
-            <ng-template #editTpl>
-              <div class="cell-wrapper">
-                <span *ngIf="clueNumbers[ri + '-' + ci]" class="clue-number">
-                  {{ clueNumbers[ri + '-' + ci] }}
-                </span>
-
-                <input
-                  maxlength="1"
-                  [value]="userGrid[ri][ci]"
-                  (input)="onInputChange($event, ri, ci)"
-                  [disabled]="!isEditableCell(ri, ci)"
-                  class="cell"
-                  [ngClass]="{
-                    correct:    cellStatus[ri][ci] === 'correct',
-                    incorrect:  cellStatus[ri][ci] === 'incorrect',
-                    unused:     !usedMask[ri][ci]
-                  }"
-                />
-              </div>
-            </ng-template>
-          </td>
-        </tr>
-      </table>
-    </div>
-
-    <div class="hints-wrapper">
-      <div class="hints">
-        <div class="hint-section">
-          <h3>Horizontales</h3>
-          <ul>
-            <li *ngFor="let h of horizontalHints">
-              {{ clueNumbers[h.x + '-' + h.y] }}. {{ h.clue }}
-            </li>
-          </ul>
-        </div>
-
-        <div class="hint-section">
-          <h3>Verticales</h3>
-          <ul>
-            <li *ngFor="let h of verticalHints">
-              {{ clueNumbers[h.x + '-' + h.y] }}. {{ h.clue }}
-            </li>
-          </ul>
-        </div>
-      </div>
-    </div>
-  </div>
-
-  <div class="controls">
-    <button class="btn-checkA" (click)="checkAnswers()">Revisar respuestas</button>
-    <button class="btn-reset" (click)="resetGame()">Reiniciar</button>
-    <button class="btn-reveal" (click)="revealLetter()" [disabled]="revealLeft === 0">
-      Revelar letra
-    </button>
-    <span class="help-counter">Ayudas: {{ revealLeft }}</span>
-    <span class="score" *ngIf="finalScore > 0">Puntaje: {{ finalScore | number:'1.0-0' }}</span>
-  </div>
-
-  <div *ngIf="isCorrect !== null" class="feedback">
-    <p *ngIf="isCorrect; else wrongTpl">✅ ¡Todas correctas!</p>
-    <ng-template #wrongTpl>❌ Hay errores. Revisa las palabras.</ng-template>
-  </div>
+<div class="crossword-placeholder">
+  <h2>Crossword</h2>
+  <p>Componente placeholder restaurado. Integraremos la lógica completa luego.</p>
 </div>
