+++ src/app/pages/games/game-sequence/game-sequence.component.ts
@@ -10,7 +10,7 @@ type Difficulty = 'facil' | 'medio' | 'dificil';
 @Component({
   selector: 'app-game-sequence',
   standalone: true,
-  imports: [CommonModule, FormsModule, NavComponent],
+  imports: [CommonModule, FormsModule, NavComponent,],
   templateUrl: './game-sequence.component.html',
   styleUrls: ['./game-sequence.component.scss']
 })
